Removed utils.go
